+++ CHANGELOG.md
@@ -24,6 +24,8 @@
   "WebSockets not enabled" path testable locally for the first time.
   ([#678](https://github.com/fastly/Viceroy/pull/678))
 
+- Future-proofed cache override hostcall ([#659](https://github.com/fastly/Viceroy/pull/659))
+
 ## 0.20.1 (2026-07-16)
 
 - Fix wiggle abi implementation for bot detection mocking. ([#664](https://github.com/fastly/Viceroy/pull/664))
@@ -32,7 +34,6 @@
 
 - Always use an `unsafe` block to `unsafe_main_ptr`. ([#643](https://github.com/fastly/Viceroy/pull/643))
 - Add support for healthchecking dynamic backends ([#635](https://github.com/fastly/Viceroy/pull/635))
-- Add support for headers-based bot detection mocking, as well as headless bots in the API, updated to v0.13.0 of Fastly crate ([#658](https://github.com/fastly/Viceroy/pull/658))
 
 ## 0.19.0 (2026-06-16)